Move Slack webhooks into .env files (not on public repository)
