+++ src/api/services/admin.py
@@ -4,7 +4,7 @@ from _main_.utils.common import serialize, serialize_all
 from api.store.admin import AdminStore
 from _main_.utils.constants import ADMIN_URL_ROOT, COMMUNITY_URL_ROOT
 from _main_.utils.emailer.send_email import send_massenergize_rich_email
-from _main_.utils.constants import SLACK_COMMUNITY_ADMINS_WEBHOOK_URL
+from _main_.settings import SLACK_COMMUNITY_ADMINS_WEBHOOK_URL
 from .utils import send_slack_message
 from typing import Tuple
